+++ changelog.md
@@ -1,8 +1,12 @@
 # Change log
 
+## v0.1.2
+
+* added `watchinnerheightchange` option. With this option, scrollfollowable watches inner height updates with timer.
+
 ## v0.1.1
 
-* added `keepholderheight` and `originalcontainerheight` option. With this option, scrollfollowable fixes container's height while the inner is fixed.
+* added `keepholderheight` and `originalcontainerheight` option. With these options, scrollfollowable fixes container's height while the inner is fixed.
 
 ## v0.1.0